% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_dictionary}
\alias{data_dictionary}
\title{Data dictionary of derived variables}
\usage{
data_dictionary()
}
\value{
data.table with \code{variable}, \code{stage}, \code{definition}.
}
\description{
Data dictionary of derived variables
}
