% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{host_id}
\alias{host_id}
\title{Identify the machine running this session}
\usage{
host_id()
}
\value{
Character scalar \code{"<hostname>:<pid>"}.
}
\description{
Used in lock owner files and logs so that a stale lock can be traced to a
host and process.
}
