% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{job_fun_default}
\alias{job_fun_default}
\title{Default job function for \code{\link[=worker]{worker()}}: SC for stage "sc", one outcome for "post"}
\usage{
job_fun_default()
}
\value{
A function \verb{(run, job, lock)}.
}
\description{
Default job function for \code{\link[=worker]{worker()}}: SC for stage "sc", one outcome for "post"
}
