% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locks.R
\name{lock_heartbeat}
\alias{lock_heartbeat}
\title{Refresh a lock's heartbeat}
\usage{
lock_heartbeat(lock)
}
\arguments{
\item{lock}{A \code{bsc_lock}.}
}
\value{
The lock, invisibly.
}
\description{
Refresh a lock's heartbeat
}
