% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locks.R
\name{lock_path}
\alias{lock_path}
\title{Path of a lock directory}
\usage{
lock_path(dir, name)
}
\arguments{
\item{dir}{Directory holding the lock.}

\item{name}{Lock name (without \code{.lock}).}
}
\value{
Character path.
}
\description{
Path of a lock directory
}
