% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{post_backend_scmbayes}
\alias{post_backend_scmbayes}
\title{scmBayesPost backend}
\usage{
post_backend_scmbayes(p, scfit, outcome, spec, lock = NULL)
}
\arguments{
\item{p}{Panel.}

\item{scfit}{Saved SC fit list.}

\item{outcome}{Outcome name.}

\item{spec}{Run spec.}

\item{lock}{Optional lock for heartbeats.}
}
\value{
list(post, gdata_light).
}
\description{
scmBayesPost backend
}
