% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{run_load}
\alias{run_load}
\title{Load a registered run by id or name}
\usage{
run_load(x, root = NULL)
}
\arguments{
\item{x}{Run id (10-character) or run name.}

\item{root}{Project root.}
}
\value{
A \code{bsc_run}.
}
\description{
Load a registered run by id or name
}
