% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_sc}
\alias{run_sc}
\title{Run the SC stage (with its own lock) and diagnostics}
\usage{
run_sc(run)
}
\arguments{
\item{run}{A \code{bsc_run}.}
}
\description{
Run the SC stage (with its own lock) and diagnostics
}
