% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.R
\name{sc_event_time_ate}
\alias{sc_event_time_ate}
\title{Event-time average effects from the SC fit}
\usage{
sc_event_time_ate(scfit)
}
\arguments{
\item{scfit}{The list saved by \code{\link[=fit_sc]{fit_sc()}}.}
}
\value{
data.table: outcome, tau (event week 0..H), mean, se, n.
}
\description{
Event-time average effects from the SC fit
}
