% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{spec_lean}
\alias{spec_lean}
\title{Lean-matching variant of the main specification}
\usage{
spec_lean(...)
}
\arguments{
\item{...}{Dotted-path overrides passed to \code{\link[=spec_modify]{spec_modify()}}.}
}
\value{
A \code{bsc_spec}.
}
\description{
Lean-matching variant of the main specification
}
