% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{spec_main}
\alias{spec_main}
\title{Working main specification}
\usage{
spec_main(...)
}
\arguments{
\item{...}{Dotted-path overrides passed to \code{\link[=spec_modify]{spec_modify()}}.}
}
\value{
A \code{bsc_spec}.
}
\description{
Working main specification
}
