% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{spec_pilot}
\alias{spec_pilot}
\title{Pilot-cohort specification}
\usage{
spec_pilot(...)
}
\arguments{
\item{...}{Dotted-path overrides passed to \code{\link[=spec_modify]{spec_modify()}}.}
}
\value{
A \code{bsc_spec}.
}
\description{
Pilot-cohort specification
}
