% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_unit_effects}
\alias{summarise_unit_effects}
\alias{summarise_ate}
\alias{summarise_gamma}
\title{Unit-level, ATE and gamma summaries}
\usage{
summarise_unit_effects(tau_draws, probs = c(0.025, 0.5, 0.975))

summarise_ate(tau_draws, probs = c(0.025, 0.5, 0.975))

summarise_gamma(post, probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{tau_draws}{Matrix from \code{\link[=extract_tau_draws]{extract_tau_draws()}}.}

\item{probs}{Quantiles. Default 2.5/50/97.5.}

\item{post}{Post result with \code{gamma_samples}.}
}
\value{
data.table.
}
\description{
Unit-level, ATE and gamma summaries
}
