% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{timeline_table}
\alias{timeline_table}
\title{Table of timeline events with both week rules}
\usage{
timeline_table(tl = NULL)
}
\arguments{
\item{tl}{A \code{bsc_timeline}; default \code{\link[=bsc_timeline]{bsc_timeline()}}.}
}
\value{
data.table with \code{event}, \code{date}, \code{weekday}, \code{wID_containing},
\code{wID_on_or_after}, \code{week_starts}.
}
\description{
Table of timeline events with both week rules
}
