% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WALLET_SHARES}
\alias{WALLET_SHARES}
\title{Wallet-share definitions: numerator (IRS sub-category) / parent category}
\format{
An object of class \code{list} of length 8.
}
\usage{
WALLET_SHARES
}
\description{
Wallet-share definitions: numerator (IRS sub-category) / parent category
}
\keyword{datasets}
