% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{week_start}
\alias{week_start}
\title{Monday on which a week index starts}
\usage{
week_start(wID, origin = NULL)
}
\arguments{
\item{wID}{Integer vector.}

\item{origin}{Date; default the timeline origin.}
}
\value{
Date vector.
}
\description{
Monday on which a week index starts
}
