% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{WALLET_SHARES}
\alias{WALLET_SHARES}
\title{Wallet-share definitions: numerator (IRS sub-category) / parent category}
\usage{
WALLET_SHARES
}
\description{
Wallet-share definitions: numerator (IRS sub-category) / parent category
}
